Add files via upload
The user can create a file if the file they are searching does not exist. The user can also search for specific words and the matched words from the content will be highlighted. The number of words will also be displayed once the file has been opened but it will not be refreshed as the user modify the content of the file. Lastly, the file will be saved together with the content, whether it is modified or not.